+++ src/main/java/com/example/basicauth/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.basicauth;
+package com.example.basicauth.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
